refactor: step3 - step 별로 구분을 하기 위해 패키지 구조 변경
