+++ src/main/java/docs/step2.md
@@ -1,4 +1,4 @@
-# 1단계 - Reflection
+# 2단계 - QueryBuilder DDL
 
 - [X] 요구사항 1 - 아래 정보를 바탕으로 create 쿼리 만들어보기
