+++ src/main/java/domain/step1/Car.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.step1;
 
 import annotation.PrintView;
