+++ src/main/java/domain/step2/Person1.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.step2;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
