+++ src/main/java/domain/step2/Person2.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.step2;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
